Update classes and add new exception
As the `detectFeelings` function from FaceAnalyzer class accepts only a
base64 string as argument its logic was modified to decode the
string into bytes, moreover, the `imageToBase64` function from
ImageHandler was modified to convert the bytes into a base64 string.

QueryHandler was updated with a new method: `getUserInfoWithID` and also
a new Exception type was added for cases that an user isn't
found in Moodle database.